Add some useful files, tweak the .gitignore
.devroot is a marker file for dev builds, so when loading the GitHub.Unity
project directly we can skip the installer logic

Add the license file to the Unity project root

Ignore the ProjectVersion.txt file. We're going to be jumping between
Unity versions, so having this file changed every time is going to get
annoying very quickly. It should contain whatever minimum version we
support, when we know what it is.