Move things around to support building with Unity and VS from the same source
The root solution loads the projects inside the src folder and subfolders, and
the GitHub.Unity project puts its output DLLS in the unity\TestProject folder,
which serves as a Unity test project for the compiled DLLs. If we need to debug
the code in Unity, the src\UnityExtension folder is also a Unity project, and
Unity compiles the sources files in it, so everything should magically work.

ignore rules for Unity and non-Unity projects are mutually exclusive
(csproj files being ignored, DLLs allowed, .meta files being included in
Unity projects), so we need to keep the Unity projects in a separate tree from
the rest of the .net projects in order to easily have gitignore rules that work
for both