Don't spam the main thread with spurious WaitForTask calls
We don't have to actively schedule a call to WaitForTask on the main thread when we're running background threads. The loop won't grab another task until the activeTask is done, so it will just sleep and check periodically. Actively spamming the main thread with repeated wait calls causes slowdowns for Unity - WaitForTask schedules itself to run, and the Tasks main loop also does, so every frame two calls were being made to it - excessive!