Fix issue where processes that error out would never finish
If a process can't find the executable to run or otherwise throws an exception on Start, then it will never run and will never "exit". The uncaught exception would just get caught by Unity and hidden away, and the tasks loop would just endlessly wait for the process to finish, and hang the whole thing.