Fix NRE, show settings when there's no repo
Move the Initialize repository UI to the history view
so that we can have that one and the settings tab available
when there's no repository yet, but only show the git
and trace logging settings in the settings view in this case.

Fix a few NRE triggered by OnGUI when there's no repository
initialized.