Refactored HookupHandlers to make use of the task system
HookupHandlers has will disable the repository watcher and toggle the busy flag by default.

Tasks that directly cause changes to the file system should toggle the stop watcher flag.
The only tasks that are called from RepositoryManager and do not directly cause changes are:
- GitStatusTask
- GitFetchTask
- GitPushTask
- GitRemoteAdd
- GitRemoteRemove
- GitRemoteChange
- GitDeleteBranch
- GitCreateBranch
- GitLockTask
- GitUnlockTask
- GitLogTask
- GitListLocksTask

Tasks that are exclusive or batches of exclusive operations should toggle the busy flag.
The only tasks that are called from RepositoryManager and are not exclusive are:
- GitLogTask
- GitListLocksTask