Fix start/end handlers and tweak Finally callback
It's useful to have the success of the state in the Finally handler that is
called in-thread (the ITask-based Finally handlers already get this information),
so add that to the callback.

In order for the API not to be confusing, make the task affinity parameter
mandatory on the Finally overloads that run on a separate thread.

Fix the HookupHandlers method to use OnStart/Finally so that the callbacks get
called on the same thread and not spin up new threads just to set flags. They're
already running on non-main threads and doing very simple things so this should
be fine.