Make sure events are unhooked when we're disposing so they don't leak
Most of our objects are singletons that live through the lifetime of the process
(the lifetime being the time between Unity reloading domains whenever it recompiles
or the user enters play mode), we never really worry much about unhooking events
to allow objects to be GC'd and to not raise events on objects that are dead.
In our tests, however, that's a problem, since we're not reloading the domain and
killing all the things between one test and the other, which means events can
definitely leak and may cause problems.

This makes sure we break event loops so that tests don't trample all over each
other. It's not a thorough audit of everything that we need to dispose, but it's
a start.