Dedupe code in ApiClient. Also, cache user data and avoid using the keychain username
The system keychain may have a valid token with an invalid username
because every git client fills out that information slightly different
and the username is not relevant for accessing the API
(only the token is needed). If we encounter a mismatch, don't throw
immediately (the token might be perfectly valid), but always doublecheck
the github user to make sure the username of the connection matches
the user that the token is associated with.

Cache the GitHubUser object in the Connection list (non serialized) so that
if we fill out once and all the information matches, we can reuse it and
avoid pinging the API every time. If another git client tramples all over
our keychain, it's highly likely that they'll also not fill out the username
information anyway, and we always grab the user from GitHub if that happens
so we should be relatively safe trusting the cached GitHubUser object
(it'll get thrown away whenever Unity reloads so it's also likely that we'll
keep refreshing it anyways)

In the process of this, cleaned up some duplicated and uneeded code that could
lead to issues later on.