Add latest version feed capabilities
Add a couple of serializable types representing information about
a package - url to download it from, version, release notes, any
other message we want to show in the UI when we inform the user of
an update.

Extend the TestApp command line tool (that we're using to test shelling out
to the command Add a tool to generate the package json for a given version and
host - we always assume that we're serving from
[host]/unity/releases/github-for-unity-[version].unitypackage.

Rename TestApp to CommandLine

Add an option to run a local web server in the CommandLine app, so
we can test update feeds and other download tasks locally.