Feed model for downloading packages; refactor installers
This adds a json format for specifying packages, with data about
where to download zips from, md5 checksums, release notes and
messages to show in the UI

Also moving the git and octorun installers away from a task-based
model, it's just too hard to maintain. Startup just runs a thread
now and all tasks are manually invoked in that thread via the Run()
and RunWithReturn() methods (i.e. they're not tasks at all)