Refactor how we run tasks synchronously so that TPL tasks aren't broken
Our task system uses TPL tasks (System.Threading.Tasks.Task) under the hood
(every ITask really is a TPL task whose body is the callback or
Run/RunWithReturn/RunWithData method override), so when we run one of our
ITask, what we're really doing is scheduling the underlying TPL task to be
executed. If we want to run the ITask synchronously, it's straightforward to
just execute the Run/RunWithReturn/RunWithData method that does the actual work.

Because of this, we also have an overload to pass in an existing TPL task into
an ITask, so that we can also run async/await task types in our threading system
(so that we can control the thread these tasks are going to be run in, given that
async/await has no threading control model).

The problem with allowing a TPL task to be set directly into an ITask is that the
TPL task wasn't set up in a way that exposed a Run/RunWithReturn/RunWithData
method that could be called synchronously.This PR moves things around so that
instead of calling directly the Run/RunWithReturn/RunWithData methods, there is one
RunSynchronously method that does the right thing for all task types, regardless of
how they're initialized.

This has the added benefit that the exact same method (`RunSynchronously`) is executed
independent of who's calling it
  - if the ITask is running on the scheduler, `RunSynchronously` is the task body that
    the scheduler will execute
  - if the user wants to run it in thread, they can call it directly.