Possible fix for the wrong git being picked up in some cases
Even though environment variable keys are case-insensitive, they're
exposed via a hashset that is case-sensitive. This means that in
some cases, when the environment variable is named `Path` for eg,
setting PATH=something causes a duplication of the keys Path/PATH,
and the actual original environment variable that the system reads
is not updated. WTF.

This change looks up environment variables by a case insensitive
search. It might fix #1044, since that problem is being caused by
the wrong git being used in some operations, so I'm going to take
a wild guess that maybe in those cases the wrong environment variable
is getting read.