Switch to using Git for Unity from https://github.com/spoiledcat/Git-for-Unity
https://github.com/spoiledcat/Git-for-Unity is now the authoritative source for
Git for Unity, so we only need to build the UI bits on top of that library.
Most of the work for Git for Unity happens at the
https://github.com/Unity-Technologies/Git-for-Unity fork, and the
spoiledcat/Git-for-Unity master repo is for hosting lfs, tracking stable
branches, and so I can actually have control over who does what on it.